Update Upcoming tasks on sideboard
Refactor createProject to now return a projectObj and a projectNode
Refactor createTask to now return a taskObj and a taskNode
This allows to access project and task objects from the code.

Projects and tasks are added to profile as they are created

updateUpcoming function with some helper functions, gets all tasks from
all projects by the user, sorts them by date and then displays them in
the DOM.

A maximum of 5 tasks can be displayed under upcoming, with the most
recent tasks that are due, being shown first(arbitrary choice).